Move snackbar to its own component
